requirements.txt
python-telegram-bot
openai
python-dotenv
requests
psycopg2-binary
flask